Split episode button from player
