+++ Pocket Casts TV App/UI/Player/EpisodePlayerView.swift
@@ -112,25 +112,3 @@ struct EpisodePlayerView: UIViewControllerRepresentable {
         return item.copy() as! AVMetadataItem
     }
 }
-
-struct EpisodePlayerButton: View {
-    let model: EpisodeRowViewModel
-    @State private var isPlaying = false
-
-    var body: some View {
-        Button {
-            isPlaying = true
-        } label: {
-            EpisodeRow(model: model, isActive: false)
-        }
-        .buttonStyle(EpisodeRowButtonStyle())
-        .fullScreenCover(isPresented: $isPlaying) {
-            EpisodePlayerView(episode: model)
-                .ignoresSafeArea()
-        }
-    }
-}
-
-#Preview {
-    EpisodePlayerButton(model: EpisodeRowViewModel(episode: MockData.makeStubEpisodes().first!, podcast: MockData.makeStubPodcasts().first!))
-}
